+++ fiddles/extjs6/fiddle-0031-ArrayStore/src/meta.js
@@ -1,0 +1,13 @@
+var meta = {
+        fiddleHeader: 'Minimal JSON / Array Store',
+        fiddleSubHeader: 'Fiddle exploring how to bind an array store to a comboBox given JSON without any keys. In other words, ' +
+        'how do you populate a store when the source json looks like this:' +
+        '<pre> ' +
+        '{ "items": [ "A", "B", "C", "D", "E" ... ] }' +
+        '</pre>' +
+        '<br />'
+    },
+    fiddleStore = null;
+
+
+