+++ fiddles/extjs6/fiddle-0031-ArrayStore/src/model/Item.js
@@ -1,0 +1,6 @@
+Ext.define('Fiddle.model.Item', {
+    extend: 'Ext.data.Model',
+    fields: [
+        'id', 'text'
+    ]
+})