+++ fiddles/extjs6/fiddle-0031-ArrayStore/src/view/View.js
@@ -1,0 +1,13 @@
+
+Ext.define('Fiddle.view.View', {
+    extend: 'Ext.panel.Panel',
+    requires: [
+        'Fiddle.view.selector.ItemSelector',
+        'Fiddle.view.Controller'
+    ],
+    controller: 'fiddlecontroller',
+    items: [{
+        xtype: 'itemselector'
+    }]
+});
+