+++ scripts/fiddle-list.sh
@@ -14,6 +14,7 @@
 # 07/26/2015 - See CHANGELOG @ 201507260420
 # 09/10/2015 - See CHANGELOG @ 201508240420
 # 12/06/2015 - See CHANGELOG @ 201511100420
+# 01/16/2015 - See CHANGELOG @ 201601100420
 # ---------------------------------------------------------------------------------------------------|
 thisFile=$(echo "$0" | sed 's/\.\///g')
 echo "${thisFile}" | awk '{print toupper($0)}'
@@ -21,7 +22,7 @@ echo "${thisFile}" | awk '{print toupper($0)}'
 (
 	if [ "$#" -gt 1 ]; then  exit 86; fi
     case $1 in
-        'angular' | 'ant' | 'compass' | 'extjs5' | 'extjs6' | 'jquery' | 'three' | 'php'| 'dojo' | 'chrome' | 'node' | 'tween' | 'bash' | 'svg')
+        'python' | 'angular' | 'ant' | 'compass' | 'extjs5' | 'extjs6' | 'jquery' | 'three' | 'php'| 'dojo' | 'chrome' | 'node' | 'tween' | 'bash' | 'svg')
             count=$(ls -1 ../fiddles/$1 | grep 'fiddle' | wc -l | sed -e 's/^[[:space:]]*//');
             echo -e ""
             echo -e "--------------------------------------------------------"
@@ -64,6 +65,7 @@ case ${_rc} in
         echo -e "\t\"extjs5\"\t\tExt JS 5 Fiddle"
         echo -e "\t\"extjs6\"\t\tExt JS 6 Fiddle"
         echo -e "\t\"php\"\t\tPHP Fiddle"
+        echo -e "\t\"python\"\tPython Fiddle"
         echo -e "\t\"jquery\"\tjQuery / Bootstrap Fiddle"
         echo -e "\t\"three\"\t\three.js / WebGl Fiddle"
         echo -e "\t\"chrome\"\tChrome Extension Fiddle"