+++ scripts/fiddle-refactor.sh
@@ -17,6 +17,7 @@
 # 07/26/2015 - See CHANGELOG @ 201507260420
 # 09/10/2015 - See CHANGELOG @ 201508240420
 # 12/06/2015 - See CHANGELOG @ 201511100420
+# 01/17/2016 - See CHANGELOG @ 201601100420
 # ---------------------------------------------------------------------------------------------------|
 
 thisFile=$(echo "$0" | sed 's/\.\///g')
@@ -100,6 +101,7 @@ case ${_rc} in
         echo -e "\t\"extjs5\"\t\tExt JS 5 Fiddle"
         echo -e "\t\"extjs6\"\t\tExt JS 6 Fiddle"
         echo -e "\t\"php\"\t\tPHP Fiddle"
+        echo -e "\t\"python\"\tPython Fiddle"
         echo -e "\t\"jquery\"\tjQuery / Bootstrap Fiddle"
         echo -e "\t\"three\"\t\tthree.js / WebGl Fiddle"
         echo -e "\t\"chrome\"\tChrome Extension Fiddle"